requirements.txt
selenium
easyocr
pillow
beautifulsoup4
schedule
